type cast size

+++ 0x1E-search_algorithms/0-linear.c
@@ -15,7 +15,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		return (-1);
 	}
-	while (i < size)
+	while (i < (int)size)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (value == array[i])
